refactor: move warp in logic to own manager from PlecementManager
