+++ src/ares/managers/placement_manager.py
@@ -1,6 +1,5 @@
 import time
 from itertools import product
-from math import ceil
 from os import getcwd, path
 from typing import (
     TYPE_CHECKING,
@@ -21,13 +20,11 @@ from cython_extensions import (
     cy_find_building_locations,
     cy_get_bounding_box,
     cy_pylon_matrix_covers,
-    cy_sorted_by_distance_to,
     cy_towards,
 )
 from loguru import logger
 from sc2.constants import ALL_GAS
 from sc2.data import Race
-from sc2.ids.ability_id import AbilityId
 from sc2.ids.unit_typeid import UnitTypeId as UnitID
 from sc2.position import Point2, Point3
 from sc2.unit import Unit
@@ -45,7 +42,6 @@ from ares.consts import (
     BuildingSize,
     ManagerName,
     ManagerRequestType,
-    UnitTreeQueryType,
 )
 from ares.dicts.structure_to_building_size import STRUCTURE_TO_BUILDING_SIZE
 from ares.managers.manager import Manager
@@ -91,8 +87,6 @@ class PlacementManager(Manager, IManagerMediator):
         UnitID.UNBUILDABLEBRICKSDESTRUCTIBLE,
         UnitID.UNBUILDABLEROCKSDESTRUCTIBLE,
     }
-    PSIONIC_MATRIX_RANGE_PYLON: float = 6.5
-    PSIONIC_MATRIX_RANGE_PRISM: float = 3.75
 
     def __init__(
         self,
@@ -126,9 +120,6 @@ class PlacementManager(Manager, IManagerMediator):
             ManagerRequestType.REQUEST_BUILDING_PLACEMENT: lambda kwargs: (
                 self.request_building_placement(**kwargs)
             ),
-            ManagerRequestType.REQUEST_WARP_IN: lambda kwargs: (
-                self.request_warp_in(**kwargs)
-            ),
         }
 
         # main dict where all data is organised
@@ -148,8 +139,6 @@ class PlacementManager(Manager, IManagerMediator):
         self.WORKER_ON_ROUTE_TIMEOUT: float = self.config[PLACEMENT][
             WORKER_ON_ROUTE_TIMEOUT
         ]
-        self.warp_in_positions: set[Point2] = set()
-        self.requested_warp_ins: list[(UnitID, Point2)] = []
 
         __ares_config_location__: str = path.realpath(
             path.join(getcwd(), path.dirname(__file__), "..")
@@ -288,106 +277,6 @@ class PlacementManager(Manager, IManagerMediator):
             avoid_creep=self.ai.race != Race.Zerg,
             include_addon=include_addon,
         )
-
-    def request_warp_in(
-        self, build_from: UnitID, unit_type: UnitID, target: Optional[Point2]
-    ) -> None:
-        """
-        Get a warp in spot closest to target
-        This is intended as a simulated alternative to example:
-        `await self.find_placement(AbilityId.WARPGATETRAIN_ZEALOT, position)`
-        So prevents making query to the game client.
-
-        Parameters
-        ----------
-        build_from
-        unit_type
-        target
-
-        Returns
-        -------
-
-        """
-        if not target:
-            target = self.ai.start_location
-
-        self.requested_warp_ins.append((build_from, unit_type, target))
-
-    async def do_warp_ins(self) -> None:
-        if not self.requested_warp_ins:
-            return
-
-        power_sources: list[Unit] = (
-            self.manager_mediator.get_own_structures_dict[UnitID.PYLON]
-            + self.manager_mediator.get_own_army_dict[UnitID.WARPPRISMPHASING]
-        )
-        if not power_sources:
-            logger.warning("Requesting warp in spot, but no power sources found")
-            return
-
-        for build_from, unit_type, target in self.requested_warp_ins:
-            size = (2, 2) if unit_type == UnitID.STALKER else (1, 1)
-            power_sources = cy_sorted_by_distance_to(power_sources, target)
-            for power_source in power_sources:
-                type_id: UnitID = power_source.type_id
-                if type_id == UnitID.PYLON:
-                    half_psionic_range = 3
-                else:
-                    half_psionic_range = ceil(self.PSIONIC_MATRIX_RANGE_PRISM / 2)
-                power_source_pos: Point2 = power_source.position
-
-                positions: list[Point2] = [
-                    Point2((power_source_pos.x + x, power_source_pos.y + y))
-                    for x in range(-half_psionic_range, half_psionic_range + 1)
-                    for y in range(-half_psionic_range, half_psionic_range + 1)
-                    if Point2((power_source_pos.x + x, power_source_pos.y + y))
-                    not in self.warp_in_positions
-                ]
-
-                in_range: list[Units] = self.manager_mediator.get_units_in_range(
-                    start_points=positions,
-                    distances=1.75,
-                    query_tree=UnitTreeQueryType.AllOwn,
-                )
-
-                for i, pos in enumerate(positions):
-                    if pos in self.warp_in_positions:
-                        continue
-
-                    if [
-                        u
-                        for u in in_range[i]
-                        if cy_distance_to_squared(u.position, pos) < 2.25
-                        and not u.is_flying
-                    ]:
-                        continue
-
-                    if cy_can_place_structure(
-                        (int(pos.x), int(pos.y)),
-                        size,
-                        self.ai.state.creep.data_numpy,
-                        self.ai.game_info.placement_grid.data_numpy,
-                        self.manager_mediator.get_ground_grid.astype(np.uint8).T,
-                        avoid_creep=True,
-                        include_addon=False,
-                    ) and cy_pylon_matrix_covers(
-                        pos,
-                        power_sources,
-                        self.ai.game_info.terrain_height.data_numpy,
-                        pylon_build_progress=1.0,
-                    ):
-                        ability: AbilityId = (
-                            AbilityId.WARPGATETRAIN_STALKER
-                            if unit_type == UnitID.STALKER
-                            else AbilityId.WARPGATETRAIN_ZEALOT
-                        )
-                        pos = await self.ai.find_placement(ability, pos)
-                        self.warp_in_positions.add(pos)
-                        build_from.warp_in(unit_type, pos)
-                        self.warp_in_positions.add(pos)
-                        if unit_type == UnitID.STALKER:
-                            for p in pos.neighbors8:
-                                self.warp_in_positions.add(p)
 
     def request_building_placement(
         self,
